requirements.txt
streamlit
opencv-python
numpy
tensorflow
python-dotenv
google-generativeai
